Moving on from get_variants.pl
I think this is basically done. In get_variants() but not
get_variants_orig() you pick up what appears to be variants at the last
base of an exon, which the old code didn’t do. (Otherwise it’s got the
same output, and runs faster.) I don’t know whether this is good or
not, but I’ve sent Kelly an ask and am going to move on for now.