bug fixin'
Turns out that if two AA substitutions happened at the same position in
consecutive genes, it would only keep the first one (mostly because I
didn’t think that would ever happen). Fixed it so now it checks that
both the gene and the substitution are identical before calling it a
duplicate and tossing it.