Removing some duplicates
Adding in stop codons means we’ll get a lot of tryptic duplicates
(where the only differences are variants that come after an inserted
stop codon so they won’t be transcribed anyway). That’s fixed now! The
only duplicate left is because of this sequence:
“EEEEEREEEEEREEEEEREEEEER” which I’m mentioning both as an example of
the fact that the duplicate is actually two identical sequences at two
different positions in the same protein and because I like to read it
out loud and pretend I’m an ambulance siren.