Adding reference mismatch check
Checks SAAVs to make sure that (a) the protein the variant occurs in
exists in the reference proteome and (b) the AA we expect to replace is
in fact that AA in the reference proteome. If not, writes message to a
reference mismatch log and doesn’t add to the variant proteome.