Alternate stop codons
Stop codons can sometimes actually translate to one of five amino acids
(EOQUW). To deal with this in SAAVs, we now replace the amino acid
directly in the protein rather than retranslating (or throwing it away
for being a mismatch between genome and proteome).  (This can’t be
avoided in other variant types, unfortunately.) Next step: when we
actually do replace a stop codon, need to add in some extra sequence
after the old stop sign, since we’re just using the old protein and
don’t have that automatic continuance